Move cache helpers to separate module
